supervisor-agent/requirements.txt
langchain-openai
python-dotenv
typing-extensions
langchain-core
langgraph
